Renamed soil.all_pools to soil.pools
