+++ docs/source/api/models/soil/pools.md
@@ -24,10 +24,10 @@ language_info:
   version: 3.11.9
 ---
 
-# API documentation for the {mod}`~virtual_ecosystem.models.soil.all_pools` module
+# API documentation for the {mod}`~virtual_ecosystem.models.soil.pools` module
 
 ```{eval-rst}
-.. automodule:: virtual_ecosystem.models.soil.all_pools
+.. automodule:: virtual_ecosystem.models.soil.pools
     :autosummary:
     :members:
 ```
