+++ virtual_ecosystem/models/soil/__init__.py
@@ -6,7 +6,7 @@ Each of the soil sub-modules has its own API reference page:
 * The :mod:`~virtual_ecosystem.models.soil.soil_model` submodule instantiates the
   SoilModel class which consolidates the functionality of the soil module into a single
   class, which the high level functions of the Virtual Ecosystem can then make use of.
-* The :mod:`~virtual_ecosystem.models.soil.all_pools` provides functionality to track
+* The :mod:`~virtual_ecosystem.models.soil.pools` provides functionality to track
   all soil pools over time.
 * The :mod:`~virtual_ecosystem.models.soil.env_factors` provides functions that capture
   the impact of environmental factors on microbial rates.
