+++ virtual_ecosystem/models/soil/soil_model.py
@@ -32,8 +32,8 @@ from virtual_ecosystem.core.core_components import CoreComponents
 from virtual_ecosystem.core.data import Data
 from virtual_ecosystem.core.exceptions import InitialisationError
 from virtual_ecosystem.core.logger import LOGGER
-from virtual_ecosystem.models.soil.all_pools import SoilPools
 from virtual_ecosystem.models.soil.constants import SoilConsts
+from virtual_ecosystem.models.soil.pools import SoilPools
 
 
 class IntegrationError(Exception):
